move huge text stories
